refactor layout code
