+++ src/layout/layout.landingPage/layout.landingPage.tsx
@@ -1,5 +1,6 @@
-import { Box, Grid, makeStyles, useMediaQuery } from '@material-ui/core';
-import {Theme} from '@material-ui/core/styles';
+import { Box, Grid, makeStyles } from '@material-ui/core';
+import HeroImageBox from './heroImg.box';
+import HeroBox from './hero.box';
 import style from './layout.landingPage.style';
 import React from 'react';
 
@@ -12,28 +13,6 @@ export interface IProps {
     expandIcon: React.ReactNode;
     detailTitle: React.ReactNode;
     detailSubtitle: React.ReactNode;
-}
-
-const getDefaultHeroBoxProps = ()=>{
-    return {
-        width:'inherit',
-        height:'100%',
-        display:'flex',
-        flexDirection:'column',
-        justifyContent:'flex-end',
-        alignItems:'flex-end',
-        ml:2,
-    }
-}
-
-const getDefaultHeroImgBoxProps = ()=>{
-    return {
-        width:'inherit',
-        display:'flex',
-        justifyContent:'center',
-        alignItems:'center',
-        mr:2,
-    }
 }
 
 const LayoutLandingPage = (props:IProps) => {
@@ -50,32 +29,17 @@ const LayoutLandingPage = (props:IProps) => {
 
     const classes = makeStyles(style)();
 
-    const downLG = useMediaQuery((theme:Theme)=>theme.breakpoints.down('sm'));
-
-    let heroImgBoxProps = getDefaultHeroImgBoxProps();
-    if(downLG){
-        heroImgBoxProps["mr"] = 0;
-        heroImgBoxProps["p"] = '5%';
-    }
-
-    let heroBoxProps = getDefaultHeroBoxProps();
-    if(downLG){
-        heroBoxProps["ml"] = 0;
-        heroBoxProps['justifyContent'] = 'center';
-        heroBoxProps['alignItems'] = 'center';
-    }
-
     return (
         <Grid container>
             {/* hero image box */}
             <Grid item xs={12} md={6}>
-                <Box {...heroImgBoxProps}>
+                <HeroImageBox>
                     <img className={classes.img} src={heroImgURL} alt='hero image' />
-                </Box>
+                </HeroImageBox>
             </Grid>
             {/* hero box */}
             <Grid item xs={12} md={6}>
-                <Box {...heroBoxProps}>
+                <HeroBox>
                     {heroTitle}
                     {heroSubtitle}
                     <Box width='auto' pt={7}>
@@ -84,7 +48,7 @@ const LayoutLandingPage = (props:IProps) => {
                     <Box width='100%' pt={7}>
                     {trending}
                     </Box>
-                </Box>
+                </HeroBox>
             </Grid>
             {/* expand icon */}
             <Grid item xs={12}>
